+++ password_Manager.py
@@ -15,20 +15,23 @@ clear = lambda: os.system('cls')
 
 
 while True:    
-    print(Fore.BLUE + 'Choose an option:\n')
+    print('Choose an option:\n')
     time.sleep(0.1)
-    print(Fore.YELLOW + '''1) Check password strength
+    print('''1) Check password strength
 2) Managed saved passwords (Coming soon)
 3) Generate a new password (Coming soon)
-4) Settings (Coming soon)''')
+4) Settings''')
 
     while True:
         if keyboard.is_pressed('1'):
-            choice = 1
+            call(['python','pStrength.py'])
+            exit()
             break
+        
+        if keyboard.is_pressed('4'):
+            call(['python','settings.py'])
+            exit()
         
     clear()
 
-    if choice == 1:
-        call(['python','pStrength.py'])
-        exit()
+