+++ Changelog.md
@@ -1,3 +1,11 @@
+V1.2.0:
+    NEW Settings: currently contains appearance settings only at the momment
+    NEW Themes: Change the way the program looks
+    UPDATE: Password strength check made more intelligent:
+        Looks for common passwords (1million of them)
+        Gives a message if you password is a common one
+    UPDATE: Start up text modified
+
 V1.1.0:
     NEW: Password strength checker
     Misc: added settings coming soon option